store: split blueprints slice into modular file structure
Splits the index.ts into slice.ts, selectors.ts, types.ts, and a
barrel index.ts to establish the domain-split file convention for
Redux slices.
